refactor(playwright-reporter-llm): extract reporter.ts into internal modules
Break the 1487-line monolithic reporter.ts into cohesive modules under
src/lib/internal/, following the repo convention established by other
packages. No behavioral changes — purely structural reorganization in
preparation for open-sourcing the package.

New internal modules:
- constants.ts: caps, allowlists, signal sets
- typeGuards.ts: asRecord, asString, asNumber, asBoolean
- zipArchive.ts: ZIP parsing utilities
- textProcessing.ts: stripAnsi, capText, content type checks
- traceTiming.ts: monotonic anchor extraction and offset math
- signalFiltering.ts: console/network signal prioritization
- testResults.ts: test title, error, step, and stdio helpers
- networkProcessing.ts: header extraction, timing, request building
- consoleProcessing.ts: console entry parsing from trace events
- traceDiagnostics.ts: trace file orchestration and attachment collection
- artifacts.ts: screenshot/video artifact extraction
- attemptBuilder.ts: timeline construction and attempt result assembly

reporter.ts is now ~180 lines containing only the LlmReporter class.
All existing tests pass unchanged with no import path modifications.
