+++ package.json
@@ -80,6 +80,6 @@
   },
   "engines": {
     "node": "24.14.1",
-    "npm": "~11.12.0"
+    "npm": "~11.11.0"
   }
 }
